+++ lib/input/reader/kafka.go
@@ -259,10 +259,10 @@ func (k *Kafka) Read() (types.Message, error) {
 	k.offset = data.Offset + 1
 	msg := types.NewMessage([][]byte{data.Value})
 
-	msg.SetMetadata("key", string(data.Key))
-	msg.SetMetadata("partition", strconv.Itoa(int(data.Partition)))
-	msg.SetMetadata("topic", data.Topic)
-	msg.SetMetadata("offset", strconv.Itoa(int(data.Offset)))
+	msg.SetMetadata("kafka_key", string(data.Key))
+	msg.SetMetadata("kafka_partition", strconv.Itoa(int(data.Partition)))
+	msg.SetMetadata("kafka_topic", data.Topic)
+	msg.SetMetadata("kafka_offset", strconv.Itoa(int(data.Offset)))
 	for _, hdr := range data.Headers {
 		msg.SetMetadata(string(hdr.Key), string(hdr.Value))
 	}