+++ lib/input/sqs.go
@@ -30,37 +30,25 @@ import (
 //------------------------------------------------------------------------------
 
 func init() {
-	Constructors["mqtt"] = TypeSpec{
-		constructor: NewMQTT,
+	Constructors["sqs"] = TypeSpec{
+		constructor: NewAmazonSQS,
 		description: `
-Subscribe to topics on MQTT brokers.
-
-### Metadata
-
-This input adds the following metadata fields to each message:
-
-` + "```" + `
-- mqtt_duplicate
-- mqtt_qos
-- mqtt_retained
-- mqtt_topic
-- mqtt_message_id
-` + "```" + `
-
-You can access these metadata fields using
-[function interpolation](../config_interpolation.md#metadata).`,
+Receive messages from an Amazon SQS URL, only the body is extracted into
+messages.`,
 	}
 }
 
 //------------------------------------------------------------------------------
 
-// NewMQTT create a new MQTT input type.
-func NewMQTT(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
-	m, err := reader.NewMQTT(conf.MQTT, log, stats)
-	if err != nil {
-		return nil, err
-	}
-	return NewReader("mqtt", reader.NewPreserver(m), log, stats)
+// NewAmazonSQS creates a new Amazon SQS input type.
+func NewAmazonSQS(conf Config, mgr types.Manager, log log.Modular, stats metrics.Type) (Type, error) {
+	return NewReader(
+		"sqs",
+		reader.NewPreserver(
+			reader.NewAmazonSQS(conf.SQS, log, stats),
+		),
+		log, stats,
+	)
 }
 
 //------------------------------------------------------------------------------